pyproject.toml
[project]
name = "ciekawa-invoices"
version = "0.1.0"
description = "Tool to automatically process invoices"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.43.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.6.0",
    "google-genai>=1.52.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
